Code refactor (no functional changes)
