+++ clients/daco-dialect-support/package.json
@@ -103,11 +103,12 @@
   },
   "scripts": {
     "vscode:prepublish": "npm run build",
+    "prebuild": "del-cli ./src/generated ./src/antlr/.antlr out dist/extension.*",
     "build": "npm run compile && esbuild src/extension.ts --bundle --sourcemap --target=node16.17.1,chrome108 --outdir=dist --external:vscode --format=cjs",
     "compile": "npm run grammar && tsc -p ./",
     "grammar": "npm run lexer && npm run parser",
-    "lexer": "antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/DaCoLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/CopybookLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/VariableLexer.g4",
-    "parser": "antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/DaCoParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/CopybookParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/VariableParser.g4",
+    "lexer": "antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/DaCoStatementsLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/ProgramLexer.g4 && antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/VariableLexer.g4",
+    "parser": "antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/DaCoStatementsParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/ProgramParser.g4 && antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/VariableParser.g4",
     "watch": "tsc -watch -p ./",
     "lint-format": "prettier -c .",
     "lint-quality": "eslint ./src",
@@ -129,6 +130,7 @@
     "@vscode/test-electron": "^2.5.2",
     "@vscode/vsce": "2.19.0",
     "antlr-ng": "^1.0.10",
+    "del-cli": "^7.0.0",
     "esbuild": "^0.28.0",
     "eslint": "^9.39.4",
     "eslint-plugin-sonarjs": "^1.0.4",
