+++ clients/daco-dialect-support/src/engine/modifiers/predefined.ts
@@ -13,7 +13,7 @@
  */
 
 import { IDocumentProcessingContext } from "@code4z/cobol-dialect-api";
-import { SettingsService } from "./services/settings";
+import { SettingsService } from "../services/settings";
 
 const SOURCE = "Generated by DaCo Dialect Support (Predefined Sections)";
