+++ clients/daco-dialect-support/src/test/unit/engine/modifiers/copyfrom.spec.ts
@@ -13,9 +13,9 @@
  */
 
 import * as vscode from "vscode";
-import { processCopyFrom } from "../../../engine/copyfrom";
-import { VariableDescriptor } from "../../../engine/parsing";
-import { createMessageService } from "./utils";
+import { processCopyFrom } from "../../../../engine/modifiers/copyfrom";
+import { VariableDescriptor } from "../../../../engine/model";
+import { createMessageService } from "../utils";
 
 describe("copy-from parsing test", () => {
   const context: any = {
